+++ CHANGELOG.md
@@ -159,6 +159,7 @@
 - PR #2275: Fix RF fit memory leak
 - PR #2274: Fix parameter name verbose to verbosity in mnmg OneHotEncoder
 - PR #2277: Updated cub repo path and branch name
+- PR #2282: Fix memory leak in Dask RF concatenation
 
 # cuML 0.13.0 (Date TBD)